improve evm-address doc snippets
